requirement.txt
crewai
firecrawl
pandas
pydantic
python-dotenv
requests
playwright
textwrap